Failed test with image maps
I thought we might be able to use image maps to put anchors inside
the image that could be linked to by a navigation bar at the top.
As far as I can tell, that doesn't work; it just focuses the image
itself.  We might have to split the final image along secion boundaries
and insert anchors between them.